Move paginate tests and unexport mock paginate function
